requirements.txt
Django==4.2.2
django-userforeignkey==0.5.0
pillow==9.4.0
djangorestframework==3.14.0
gunicorn==20.1.0
django-debug-toolbar==4.1.0
djoser==2.2.0
drf-extensions==0.7.1